Separate spgrep.irreps
