+++ src/spgrep/corep.py
@@ -7,17 +7,14 @@ from typing import Literal
 import numpy as np
 
 from spgrep._constants import ATOL, MAX_NUM_RANDOM_GENERATIONS, RTOL
-from spgrep.irreps import (
-    enumerate_small_representations,
-    enumerate_unitary_irreps,
-    is_equivalent_irrep,
-)
 from spgrep.rep.group import get_inverse_index
+from spgrep.rep.irreps import is_equivalent_irrep
 from spgrep.rep.representation import get_character, get_intertwiner
 from spgrep.spinor import (
     enumerate_spinor_small_representations,
     get_spinor_unitary_rotation,
 )
+from spgrep.symmetry.enumerate import enumerate_small_representations, enumerate_unitary_irreps
 from spgrep.symmetry.group import (
     decompose_by_maximal_space_subgroup,
     get_cayley_table,
