+++ tests/rep/test_rep_pir.py
@@ -1,0 +1,66 @@
+from __future__ import annotations
+
+import numpy as np
+
+from spgrep.rep.pir import get_physically_irrep
+
+
+def test_physically_irrep():
+    irrep = np.array(
+        [
+            [
+                [1.00000000e00 + 0.00000000e00j, -6.93889390e-17 - 5.55111512e-17j],
+                [-6.93889390e-17 + 5.55111512e-17j, 1.00000000e00 + 0.00000000e00j],
+            ],
+            [
+                [1.75723416e-02 + 2.77555756e-17j, -9.14828354e-01 - 4.03460401e-01j],
+                [-9.14828354e-01 + 4.03460401e-01j, -1.75723416e-02 + 2.77555756e-17j],
+            ],
+            [
+                [-1.00000000e00 + 2.77555756e-17j, 5.55111512e-17 - 5.55111512e-17j],
+                [3.46944695e-17 + 0.00000000e00j, -1.00000000e00 + 1.38777878e-17j],
+            ],
+            [
+                [-5.00000000e-01 + 3.71108152e-01j, -7.18554655e-01 - 3.09770799e-01j],
+                [7.18554655e-01 - 3.09770799e-01j, -5.00000000e-01 - 3.71108152e-01j],
+            ],
+            [
+                [-7.73548252e-01 + 0.00000000e00j, -6.19768309e-01 + 1.32326661e-01j],
+                [-6.19768309e-01 - 1.32326661e-01j, 7.73548252e-01 + 0.00000000e00j],
+            ],
+            [
+                [7.91120593e-01 + 0.00000000e00j, -2.95060045e-01 - 5.35787062e-01j],
+                [-2.95060045e-01 + 5.35787062e-01j, -7.91120593e-01 + 8.67361738e-19j],
+            ],
+            [
+                [-5.00000000e-01 - 3.71108152e-01j, 7.18554655e-01 + 3.09770799e-01j],
+                [-7.18554655e-01 + 3.09770799e-01j, -5.00000000e-01 + 3.71108152e-01j],
+            ],
+            [
+                [5.00000000e-01 + 3.71108152e-01j, -7.18554655e-01 - 3.09770799e-01j],
+                [7.18554655e-01 - 3.09770799e-01j, 5.00000000e-01 - 3.71108152e-01j],
+            ],
+            [
+                [-7.91120593e-01 + 0.00000000e00j, 2.95060045e-01 + 5.35787062e-01j],
+                [2.95060045e-01 - 5.35787062e-01j, 7.91120593e-01 + 1.04083409e-17j],
+            ],
+            [
+                [7.73548252e-01 - 5.55111512e-17j, 6.19768309e-01 - 1.32326661e-01j],
+                [6.19768309e-01 + 1.32326661e-01j, -7.73548252e-01 + 0.00000000e00j],
+            ],
+            [
+                [5.00000000e-01 - 3.71108152e-01j, 7.18554655e-01 + 3.09770799e-01j],
+                [-7.18554655e-01 + 3.09770799e-01j, 5.00000000e-01 + 3.71108152e-01j],
+            ],
+            [
+                [-1.75723416e-02 + 0.00000000e00j, 9.14828354e-01 + 4.03460401e-01j],
+                [9.14828354e-01 - 4.03460401e-01j, 1.75723416e-02 - 2.77555756e-17j],
+            ],
+        ]
+    )
+
+    physically_irrep = get_physically_irrep(irrep, indicator=1)
+    assert np.allclose(
+        np.abs(np.linalg.det(physically_irrep)),
+        1.0,
+    )
