+++ tests/symmetry/test_symmetry_pir.py
